+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All Notable changes to `sms` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## Date - 2018-03-09
+
+### Fixed
+- Array Merge Issue for Textlocal & LinkMobility.
 
 ### Added
 - [Twilio](https://www.twilio.com) Driver added.